+++ src/modules/mo/mo_dhl/Adapter/DHLAdapter.php
@@ -213,4 +213,12 @@ class DHLAdapter
     {
         return $this->getSdk()->buildGKV();
     }
+
+    /**
+     * @return \Mediaopt\DHL\Api\Retoure
+     */
+    public function buildRetoure()
+    {
+        return $this->getSdk()->buildRetoure();
+    }
 }