+++ src/modules/mo/mo_dhl/Api/Credentials.php
@@ -141,8 +141,21 @@ class Credentials
         return new static('https://cig.dhl.de/services/production/soap', $username, $password, $ekp, false);
     }
 
-    public static function createCustomerGKVCredentials($username, $password)
+    /**
+     * @param string $username
+     * @param string $password
+     * @return static
+     */
+    public static function createCustomerCredentials($username, $password)
     {
         return new static('', $username, $password);
+    }
+
+    /**
+     * @return string
+     */
+    public function getBasicAuth()
+    {
+        return base64_encode($this->getUsername() . ':' . $this->getPassword());
     }
 }